backend: Replace recTypes by single recType
The recommended node for recType="next" is given by the path attribute,
not by a special node type.
This is possible since the first node of "path" is always "next" and
hence "next" offers no additional information over "path". As such, it
does not make sense to allow giving both options at the same time or to
represent the output of "next" and "path" in different ways.